requirements.txt
aiogram==3.7.0
yt-dlp
requests
beautifulsoup4
python-dotenv
moviepy